# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API Restaurante
  version: 1.0.0
  description: API para gestión de platos, pedidos y usuarios
servers:
  - url: http://localhost:3000
paths:
...
  /api/usuarios:
    get:
      summary: Obtener todos los usuarios
      responses:
        '200':
          description: Lista de usuarios

    post:
      summary: Crear un usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                correo:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Usuario creado
  /api/usuarios/{id}:
    get:
      summary: Obtener un usuario por id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario encontrado
        '404':
          description: Usuario no encontrado

    put:
      summary: Actualizar un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                correo:
                  type: string
      responses:
        '200':
          description: Usuario actualizado

    delete:
      summary: Eliminar un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Usuario eliminado
        '404':
          description: Usuario no encontrado